Move render stage, shader map, uniforms map to frontend. Set up shared render resources for stages.

+++ bullet_hell/src/graphics/backend/opengl/render_buffers.cpp
@@ -269,7 +269,7 @@ void RenderBuffers::load_bones_matrices_buffer(const ModelList& models)
 						}
 					}
 				}
-				current_offset += frame.bone_matrices.size();
+				current_offset += static_cast<int>(frame.bone_matrices.size());
 			}
 		}
 	}
